+++ CHANGELOG.md
@@ -6,7 +6,7 @@ project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-## [0.2.6] - 2026-09-06
+## [0.2.6] - 2026-09-07
 
 ### Added